reorganize the structure
